apps/api/requirements.txt
# AdVantage API v3.0.0

# Core
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.10.0
pydantic-settings>=2.7.0

# Database
motor>=3.6.0
pymongo>=4.10.0

# Auth
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.20

# Utilities
python-dateutil>=2.9.0
anyio>=4.8.0

# Testing
pytest>=8.3.0
pytest-asyncio>=0.25.0
httpx>=0.28.0